refactor(storefront): unify Saleor-only loaders on createServiceLoader
