+++ apps/storefront/src/services/lazy-loaders/cart.ts
@@ -1,9 +1,9 @@
 import type { Logger } from "@nimara/infrastructure/logging/types";
 
 import { clientEnvs } from "@/envs/client";
-import { createServiceLoader } from "@/services/integrations/create-loader";
+import { createServiceLoader } from "@/services/utils/create-loader";
 
-import { emptyCartService, isSaleorConfigured } from "./empty-services";
+import { emptyCartService, isSaleorConfigured } from "../utils/empty-services";
 import { getRequiredSaleorApiUrl } from "./required-env";
 
 /**
