+++ apps/storefront/src/services/lazy-loaders/checkout.ts
@@ -1,9 +1,12 @@
 import type { Logger } from "@nimara/infrastructure/logging/types";
 
 import { clientEnvs } from "@/envs/client";
-import { createServiceLoader } from "@/services/integrations/create-loader";
+import { createServiceLoader } from "@/services/utils/create-loader";
 
-import { emptyCheckoutService, isSaleorConfigured } from "./empty-services";
+import {
+  emptyCheckoutService,
+  isSaleorConfigured,
+} from "../utils/empty-services";
 import { getRequiredSaleorApiUrl } from "./required-env";
 
 /**
