+++ apps/storefront/src/services/lazy-loaders/user.ts
@@ -1,8 +1,8 @@
 import type { Logger } from "@nimara/infrastructure/logging/types";
 
-import { createServiceLoader } from "@/services/integrations/create-loader";
+import { createServiceLoader } from "@/services/utils/create-loader";
 
-import { emptyUserService, isSaleorConfigured } from "./empty-services";
+import { emptyUserService, isSaleorConfigured } from "../utils/empty-services";
 import { getRequiredSaleorApiUrl } from "./required-env";
 
 /**
